refactor: RequestLine class 분리
RequestHeaders의 적절하지 않은 책임 분리
